fix camera selection
Camera-selector and Barcode have been refactored so that 1.  now all available cameras are shown, 2. the selected camera is actually used by the scanner, 3. if the stored camera is not available, the error is handled and the selector shows 'Default Device' (the scanner uses automatically the first available device). Also the reset of the scanner between enhanced so that this only occurs when a page with the scanner is left